Add docs badge to top of README
Finding myself scrolling around the readme for some seconds trying to find the link to the docs, which is a couple of sections down from the top. In order to help people (and myself) to find the reference faster, I added a badge in the top that links to the book/docs.